Create an app which will have a Table view, and a navigation controller and a bar button item on the controller.
Create an app which will have a Table view, and a navigation controller and a bar button item on the controller. When you click on the bar button item it will open an alert controller. You will collect Name, Phone number from the user and then add this on the4 table view.